pyproject.toml
[project]
name = "nonebot-plugin-pxchat"
version = "0.1.0"
description = "基于AI大模型的聊天机器人，支持大模型api任意切换、上下文记忆、群聊智能参与、图片识别、MCP等功能"
readme = "README.md"
authors = [{ name = "whopxx", email = "2074447320@qq.com" }]
urls = { Repository = "https://github.com/whopxxx/nonebot-plugin-pxchat" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.16.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2>=2.4.3",
    "openai>=2.2.0",
]

[tool.nonebot]
plugins = ["nonebot_plugin_pxchat"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "nonebot2[fastapi]>=2.4.3",
]
